refactor alert into alertcontainer fixed bug when removing an alert with index as key removing two notifications at once
