+++ src/components/molecules/AlertContainer.vue
@@ -6,11 +6,22 @@
 		}"
 	>
 		<transition-group :name="transition">
-			<Alert
-				v-for="(notification, index) in notifierItems"
-				:key="index"
-				:notification="notification"
-			/>
+			<v-alert
+				v-for="notification in notifierItems"
+				:key="notification.id"
+				:type="notification.status"
+				:icon="statusIcons[notification.status]"
+				class="alert"
+				closable
+				max-width="400"
+				min-width="200"
+				border="start"
+				@click:close="removeNotifier(notification.id)"
+			>
+				<div class="alert-text mr-2" data-testId="alert-text">
+					{{ notification.text }}
+				</div>
+			</v-alert>
 		</transition-group>
 	</div>
 </template>
@@ -18,16 +29,30 @@
 <script setup lang="ts">
 import { computed } from "vue";
 import { useDisplay } from "vuetify";
-import Alert from "./Alert.vue";
-import { useNotificationStore } from "@data-app";
+import { AlertStatus, useNotificationStore } from "@data-app";
 import { storeToRefs } from "pinia";
+import {
+	mdiAlert,
+	mdiAlertCircle,
+	mdiCheckCircle,
+	mdiInformation,
+} from "@icons/material";
 
 const { notifierItems } = storeToRefs(useNotificationStore());
+const { removeNotifier } = useNotificationStore();
+
 const { xs: isMobile } = useDisplay();
 
 const transition = computed(() =>
 	isMobile.value ? "scale-transition" : "scroll-x-reverse-transition"
 );
+
+const statusIcons: { [status in AlertStatus]: string } = {
+	success: mdiCheckCircle,
+	warning: mdiAlert,
+	error: mdiAlertCircle,
+	info: mdiInformation,
+};
 </script>
 
 <style lang="scss" scoped>
@@ -45,5 +70,16 @@ const transition = computed(() =>
 	left: 0;
 	z-index: 50;
 	overflow: visible;
+}
+
+.alert {
+	margin: 0 12px 12px 0;
+	overflow: hidden;
+	background-color: rgba(var(--v-theme-white)) !important;
+}
+
+:deep(.v-btn__content .v-icon),
+.alert-text {
+	color: rgba(var(--v-theme-on-background)) !important;
 }
 </style>
